# api-swagger/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: MusicRoomAPI
host: localhost:3001
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/api/user/update-data':
    x-swagger-router-controller: user
    post:
      tags:
        - User
      # description: Route to update simple user data
      operationId: updateData
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateUserDataInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  '/api/mtv/events':
    x-swagger-router-controller: mtv
    get:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: getEvents
      parameters:
        - in: query
          name: location
          type: string
          description: latitude and longitude seperated by an escape
          required: true
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Events'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/mtv/event':
    x-swagger-router-controller: mtv
    post:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: createEvent
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateEventInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Events'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: deleteEvent
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/DeleteEventInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Events'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    patch:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: updateData
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateEventDataInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Events'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/mtv/tracks':
    x-swagger-router-controller: mtv
    get:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: getTracks
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: query
          name: playlistId
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/TracksMTV'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/mtv/track':
    x-swagger-router-controller: mtv
    post:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: addTrack
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AddTrackInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/TracksMTV'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: removeTrack
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/DeleteTrackInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/TracksMTV'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/mtv/track/vote':
    x-swagger-router-controller: mtv
    post:
      tags:
        - Music Track Vote
      # description: Route to update simple user data
      operationId: voteTrack
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/VoteTrackInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/TracksMTV'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  '/api/mpe/tracks':
    x-swagger-router-controller: mpe
    get:
      tags:
        - Music Playlist Editor
      # description: Route to update simple user data
      operationId: getPlaylistTracks
      parameters:
        - in: header
          name: X-SessionID
          type: string
          description: your sessionId here
          required: true
        - in: query
          name: playlistId
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/TracksMPE'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  UpdateUserDataInput:
    required:
      - toChange
      - newValue
    properties:
      toChange:
        type: string
      newValue:
        type: string
...
  User:
    properties:
      email:
        type: string
      name:
        type: string
      firstname:
        type: string
      avatarUri:
        type: string
      deezer:
        type: boolean
      facebook:
        type: boolean
      google:
        type: boolean
      tags:
        type: array
        items:
          type: string
      privacy:
        type: object
  Events:
    type: array
    items:
      $ref: '#/definitions/Event'
  Event:
    properties:
      id:
        type: string
      name:
        type: string
      canEdit:
        type: boolean
      playlistId:
        type: number
      restriction:
        $ref: '#/definitions/Restriction'
      allowedUsers:
        type: array
        items:
          type: string
      privacy:
        type: string
  Restriction:
    properties:
      isRestricted:
        type: boolean
      location:
        type: string
      maxDistance:
        type: integer
      startDate:
        type: integer
      endDate:
        type: integer
  CreateEventInput:
    required:
      - location
      - name
    properties:
      location:
        type: string
      name:
        type: string
  DeleteEventInput:
    required:
      - location
      - eventId
    properties:
      location:
        type: string
      eventId:
        type: string
  UpdateEventDataInput:
    required:
      - toChange
      - newValue
      - eventId
      - location
    properties:
      toChange:
        type: string
      newValue:
        type: string
      eventId:
        type: string
      location:
        type: string
  TrackMPE:
    properties:
      id:
        type: number
      albumCover:
        type: string
      artistName:
        type: string
      previewUrl:
        type: string
      title:
        type: string
  TrackMTV:
    properties:
      id:
        type: number
      albumCover:
        type: string
      artistName:
        type: string
      previewUrl:
        type: string
      title:
        type: string
      numberOfVote:
        type: number
  TracksMPE:
    type: array
    items:
      $ref: '#/definitions/TrackMPE'
  TracksMTV:
    type: array
    items:
      $ref: '#/definitions/TrackMTV'
  AddTrackInput:
    required:
      - playlistId
      - trackId
    properties:
      playlistId:
        type: number
      trackId:
        type: number
  DeleteTrackInput:
    required:
      - playlistId
      - trackId
    properties:
      playlistId:
        type: number
      trackId:
        type: number
  VoteTrackInput:
    required:
      - eventId
      - trackId
      - value
    properties:
      eventId:
        type: string
      trackId:
        type: number
      value:
        type: integer
...
  ErrorResponse:
    required:
      - error
    properties:
      error:
        type: string